refactor: split renderer host contract into shared/host-api.ts (Stage A Move 3)
Extracts the 21 ElectronAPI sub-interfaces + the Window declaration
from shared/types.ts (1338 → 828 lines) into shared/host-api.ts (542
lines). shared/types.ts re-exports everything from host-api for
backward compat, so no existing import paths break.

Why: the host contract is the part that DIFFERS between an Electron
build and a future Chrome-extension build. Keeping it separate from
wire/domain types means a future extension shim can live alongside
this file (shared/host-api-extension.ts or similar) without forking
the whole types module.

What stays where:
- shared/types.ts (lines 1-822) — wire protocol (DuoRequest/Response,
  DuoCommandName), IPC channel constants, and the data types used by
  both wire and host surfaces (BrowserState, SessionState, FileEntry,
  selection snapshots, theme state, etc.)
- shared/host-api.ts (extracted) — 21 ElectronAPI sub-interfaces
  (ElectronEnv, ElectronPtyAPI, ElectronBrowserAPI, ...) plus their
  embedded data types (FileWriteResult, HtmlFileMeta, ForwardedKeyEvent,
  ClaudePresenceState, HistorySuggestion, NewTabRequest/Result,
  InstallStatus, CliInstallStatus, PrimingInstallStatus, InstallResult,
  CanvasAction, AddToShellPathResult, ExternalRedirectedPush,
  UpdateCheckResult), the umbrella ElectronAPI, and the
  `declare global { interface Window { electron: ElectronAPI } }`
  augmentation.

Verified: npm run typecheck and npm run build both clean.
